third_party/WebKit/LayoutTests/external/wpt/NOT_FOR_LANDING/xlink.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     xmlns:xlink="http://www.w3.org/1999/xlink">
  <script xlink:href="../../../external/wpt/resources/testharness.js"/>
  <script xlink:href="../../../external/wpt/resources/testharnessreport.js"/>
  <rect width="100" height="100" fill="blue"/>
  <script>
    test(function() {
      assert_true(true, 'anything');
      assert_true(false, 'nothing');
    }, 'foo');
  </script>
</svg>
